Add command line arguments and allow them changing log behaviour.
Client and server now log to standard output with level `WARNING` and to a log file
using `SessionRotatingFileHandler` with level `DEBUG`. The default log file location is
supplied by `appdirs.user_log_dir`. A different log level as well as a different log file
location can be specified through the command line. It is also possible to disable logging
to a log file.

* python/ectec-gui/setup.cfg (options.install_requires): Add appdirs.

* python/ectec-gui/src/ectecgui/__init__.py : Add constants `APPNAME` and `APPAUTHOR` that
	can be passed to `appdirs` to get a default directory.

* python/ectec-gui/src/ectecgui/client/__main__.py : Implement logging behaviour and argparse.
* python/ectec-gui/src/ectecgui/server/__main__.py : Implement logging behaviour and argparse.